requirements.txt
annotated-types
anyio
arrow
asgiref
asttokens
async-lru
attrs
Babel
beautifulsoup4
bleach
blobfile
catboost
contourpy
cycler
Django
djangorestframework
dnspython
et-xmlfile
executing
fastjsonschema
fire
fonttools
fqdn
fsspec
geopandas
graphviz
h11
httpcore
httpx
huggingface-hub
hupper
idna
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter_client
jupyter-console
jupyter_core
jupyter-events
jupyter-lsp
jupyter_server
jupyter_server_terminals
jupyter-ui-poll
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
kneed
lightgbm
llama_models
llama_stack
lxml
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
mpmath
msal
mysql-connector-python
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
openpyxl
overrides
packaging
pandas
pandocfilters
parso
PasteDeploy
pbr
Pillow
plaster
plaster-pastedeploy
plotly
powerbiclient
prometheus_client
prompt_toolkit
psutil
pure-eval
py4j
pydantic
pydantic_core
Pygments
PyJWT
pymongo
pyogrio
pyparsing
pyproj
pyramid
pyspark
python-dateutil
python-dotenv
python-json-logger
pytz
qtconsole
QtPy
rapids
referencing
regex
requests
rfc3339-validator
rfc3986-validator
rich
rpds-py
scikit-learn
scipy
seaborn
Send2Trash
shapely
six
sniffio
soupsieve
sqlparse
stack-data
stevedore
style
sympy
tenacity
termcolor
terminado
threadpoolctl
tiktoken
tinycss2
torch
torchaudio
torchvision
tornado
tqdm
traitlets
translationstring
types-python-dateutil
typing_extensions
uri-template
urllib3
venusian
wcwidth
webcolors
webencodings
WebOb
websocket-client
widgetsnbextension
xgboost
zope.deprecation
zope.interface
